tests/requirements.txt
# Test-specific dependencies
psycopg2-binary>=2.9.9
pytest>=7.4.0
pyyaml>=6.0.1
tabulate>=0.9.0

